refactor(payment-attempt):split the transaction instead of one giant transaction

+++ payment-service/src/main/java/com/nexupay/payment/bank/dto/BankRequest.java
@@ -4,20 +4,20 @@ import java.math.BigDecimal;
 
 public class BankRequest {
     private String transactionId;
-    private Long amount;
+    private BigDecimal amount;
     private String upiId;
 
     public String getTransactionId() {
         return transactionId;
     }
 
-    public BankRequest(String transactionId, Long amount, String upiId) {
+    public BankRequest(String transactionId, BigDecimal amount, String upiId) {
         this.transactionId = transactionId;
         this.amount = amount;
         this.upiId = upiId;
     }
 
-    public Long getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
