+++ payment-service/src/main/java/com/nexupay/payment/payment/dto/request/PaymentRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public class PaymentRequest {
     @NotBlank(message = "Merchant order id is required")
     private String merchantOrderId;
@@ -12,7 +14,7 @@ public class PaymentRequest {
     private String customerPhone;
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be greater than zero")
-    private Long amount;
+    private BigDecimal amount;
     @NotBlank(message = "Currency is required")
     private String currency;
 
@@ -49,11 +51,11 @@ public class PaymentRequest {
         this.customerPhone = customerPhone;
     }
 
-    public Long getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Long amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
